% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callWrappers.R
\name{i.companies}
\alias{i.companies}
\title{Loads Information about companies}
\usage{
i.companies(tickers = NULL, ...)
}
\arguments{
\item{tickers}{A character vector of stock symbols to load detailed data on.
Loads an entire list of companies (master data feed) with compacted response values if \code{NULL}.}

\item{...}{other arguments to pass to \code{MoreArgs} of \code{intrCallMap}}
}
\value{
Data in specified format. See \code{\link{intrOptions}} for details
}
\description{
Loads Information about companies
}
\details{
This query will return companies available at intrinio with detailed information,
such as hq address, phone number, industry sector, SEC cik number etc.
}
\section{Data Feed}{

{https://intrinio.com/data/company-financials}{US Public Company Financials}
}
\examples{
\dontrun{
i.companies()
i.companies(c('MSFT', 'AAPL'))
}
}
\seealso{
\href{http://docs.intrinio.com/#companies}{Companies API documentation}.
}

