% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callWrappers.R
\name{i.dataPoint}
\alias{i.dataPoint}
\title{Most recent data point}
\usage{
i.dataPoint(tickers, tags, ...)
}
\arguments{
\item{tickers}{A character vector of stock symbols (or other eligible identifiers, such as CIK number)
to load detailed data on.}

\item{tags}{A character vector of stock symbols to load detailed data on.
Entire tag list is available \href{http://docs.intrinio.com/tags/intrinio-public#data-point}{here}}

\item{...}{other arguments to pass to \code{MoreArgs} of \code{intrCallMap}}
}
\value{
Data in specified format. See \code{\link{intrOptions}} for details
}
\description{
Returns the most recent data point (such as P/E, price, financial) for selected identifier(s)
(ticker symbol, stock market index symbol, CIK ID, etc.) for selected tag(s).
}
\details{
API allows requesting up to 150 identifier/tah combinations at a time.
E.g. 1 identifier and 150 items, or 150 identifiers and 1 tag, or 12 identifiers and 12 tags (144 combinations).
This function automatically makes multiple requests if API limits are not met, attempting to minimize number of calls.
It is assumed that you need \emph{all} data tags specified for every ticker specified,
so do not pass duplicate tickers or tags.
}
\section{Data Feed}{

\href{https://intrinio.com/data/company-financials}{US Public Company Financials}
}
\seealso{
\href{http://docs.intrinio.com/#data-point}{Data Point documentation}.
}

