% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callWrappers.R
\name{i.historicalData}
\alias{i.historicalData}
\title{Historical data}
\usage{
i.historicalData(tickers, tags, from = NULL, to = NULL, freq = NULL,
  type = NULL, ...)
}
\arguments{
\item{tickers}{tickers A character vector of stock symbols.
For non-US tickers it is possible to pass specify exchange after tick symbol, separated by colon, i.e. "TICKER:EXCHANGE"}

\item{tags}{specified standardized tag requested.
Entire list of tags is available \href{http://docs.intrinio.com/tags/intrinio-public#historical-data}{here}}

\item{from}{History start date (either Date or a character in 'YYYY-MM-DD' format),
If \code{NULL}, date would not be restricted by a minimum.
A vector of dates of the same length as tickers is also possible in order to have specific start date for each ticker.}

\item{to}{History end date (either Date or a character in 'YYYY-MM-DD' format),
If \code{NULL}, date would not be restricted by a maximum.
A vector of dates of the same length as tickers is also possible in order to have specific end date for each ticker.}

\item{freq}{Data periodicity. A string containing any of
\code{'daily', 'weekly', 'monthly', 'quarterly', 'yearly'} or \code{NULL}.
If \code{NULL} will preserve default intrinio behavior (daily).}

\item{type}{Meaningful for financial Statements data only.
The type of financial statement period. Possible values:\cr
\describe{
  \item{\code{'FY'}}{Fiscal year}
  \item{\code{'QTR'}}{Quarter}
  \item{\code{'YTD'}}{Year-to-date}
  \item{\code{'TTM'}}{Trailing twelve months}
}\cr
\code{NULL} will preserve default intrinio behavior, which is TTM for income statement and
cash flow statement and QTR for balance sheet.}

\item{...}{other arguments to pass to \code{MoreArgs} of \code{intrCallMap}}
}
\value{
Data in specified format. See \code{\link{intrOptions}} for details
}
\description{
Historical data
}
\details{
Will download history for all tickers passed for each tag passed.
Please use \code{intrCallMap} to load specific set of tags for the respective ticker.
}
\note{
Financials will not work with freq defined, so it is generally not advised to mix market data and financials.
}
\section{Data Feed}{

\href{https://intrinio.com/data/company-financials}{US Public Company Financials}
}

