% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callWrappers.R
\name{i.indices}
\alias{i.indices}
\title{Return available indices}
\usage{
i.indices(tickers = NULL, type = NULL, ...)
}
\arguments{
\item{tickers}{A character vector of stock symbols to load detailed data on.
Loads an entire list of indices (master data feed) with compacted response values if \code{NULL}.}

\item{type}{A character vector with index types.
Either \code{stock_market}, \code{economic} or \code{sic}}
}
\value{
Data in specified format. See \code{\link{intrOptions}} for details
}
\description{
Return available indices
}
\details{
This query will return indices available at intrinio with detailed information,
such as country, continent, full name and type.
}
\examples{
\dontrun{
i.indices() # over 1000 API calls
i.indices(ticker = c('$SPX', '$SSEC'))
i.indices(type = 'stock_market')
}
}
\seealso{
\href{http://docs.intrinio.com/#indices58}{Indices API documentation}.
}

