% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callWrappers.R
\name{i.prices}
\alias{i.prices}
\title{Requests market data history (OHLC, volumes, dividends, etc.)}
\usage{
i.prices(tickers, from = NULL, to = NULL, freq = "daily", ...)
}
\arguments{
\item{tickers}{tickers A character vector of stock symbols.
For non-US tickers it is possible to specify exchange after tick symbol, separated by colon, i.e. "TICKER:EXCHANGE"}

\item{from}{History start date (either Date or a character in 'YYYY-MM-DD' format),
If \code{NULL}, date would not be restricted by a minimum.
A vector of dates of the same length as tickers is also possible in order to have specific start date for each ticker.}

\item{to}{History end date (either Date or a character in 'YYYY-MM-DD' format),
If \code{NULL}, date would not be restricted by a maximum.
A vector of dates of the same length as tickers is also possible in order to have specific end date for each ticker.}

\item{freq}{Data periodicity. A string containing any of
\code{'daily', 'weekly', 'monthly', 'quarterly', 'yearly'} or \code{NULL}. \code{NULL} will preserve default intrinio behavior (daily).}
}
\value{
Data in specified format. See \code{\link{intrOptions}} for details. Columns: \cr
\describe{
	\item{\code{ticker}}{Security identifier}
	\item{\code{date}}{Price date}
	\item{\code{open}}{The actual observed first traded stock price on the trading date}
	\item{\code{high}}{The actual observed highest traded stock price on the trading date}
	\item{\code{low}}{The actual observed lowest traded stock price on the trading date}
	\item{\code{close}}{The actual observed last trade stock price on the trading date}
	\item{\code{volume}}{The actual observed number of shares of stock traded between market participants on the trading date}
	\item{\code{ex_dividend}}{The non-split adjusted dividend per share on the ex-dividend date - not available on index historical prices}
	\item{\code{split_ratio}}{The split factor on the split date - not available on index historical prices}
	\item{\code{adj_open}}{The dividend and split adjusted open price - not available on index historical prices}
	\item{\code{adj_high}}{The dividend and split adjusted high price - not available on index historical prices}
	\item{\code{adj_low}}{The dividend and split adjusted low price - not available on index historical prices}
	\item{\code{adj_close}}{The dividend and split adjusted close price - not available on index historical prices}
	\item{\code{adj_volume}}{The dividend and split adjusted volume - not available on index historical prices}}
}
\description{
Using \code{prices} endpoint downloads data for a subset of
tags, allowed by \code{\link{i.historicalData}}. However, it loads multiple
tags with a single call, which allows spending less API credits.
}
