% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intrinio.R
\name{intrAuth}
\alias{intrAuth}
\title{Authorize Web API access}
\usage{
intrAuth(key)
}
\arguments{
\item{Key}{A string containing \code{API_KEY}}
}
\value{
a list with credentials if key is not passed.
}
\description{
The Intrinio API either passes key with link over HTTPS or
via Authorization header. The latter is used in this package.
You can find your \code{API_KEY} in your user account on the website.\cr
\code{intrAuth} simply stores your credentials in \code{intrinio_globals} environment for future use.
Credentials need to be specified on every package reload.
}
\examples{
intrAuth('a543b029ec335ddfw66dd95bfa1ea3ac')
intrAuth()
}
\seealso{
\href{http://blog.intrinio.com/getting-started-with-intrinio/}{Getting Started article}
}
