% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intrinio.R
\name{intrCall}
\alias{intrCall}
\title{Generic Intrinio API call with multi-page results}
\usage{
intrCall(endpoint, pageSize = "auto", startPage = 1, endPage = NULL,
  pageRecover = TRUE, outFormat = intrOptions()$outFormat, ...)
}
\arguments{
\item{endpoint}{Name of an endpoint, such as 'companies', 'securities', 'indices', 'owners' or 'stock_exchanges'.
See \href{docs.intrinio.com}{API documentation}. May have a composite name, e.g. \code{'fundamentals/standardized'}}

\item{startPage}{A page to start loading from.}

\item{...}{Any other named flags/queries to pass.}

\item{page_size}{Page size for results. If set to \code{'auto'} will use the largest size allowed for the endpoint.
See \href{http://docs.intrinio.com/#paging-limits}{Paging Limits} or \code{intrOptions()$maxPageSize}}
}
\value{
a list with data.
}
\description{
Unless API call results fit into Intrinio limit, they are split into multiple pages.
This function requests each page separately and combines results.
}
\examples{
intrCall('companies') # Pulls Master Data Feed (first page).
intrCall('companies', identifier = 'AAPL')
intrCall('fundamentals/standardized', identifier = 'AAPL', statement = 'income_statement', type = 'FY')
}
