% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intrinio.R
\name{intrCallMap}
\alias{intrCallMap}
\title{A multi-request wrapper around intrCall}
\usage{
intrCallMap(endpoint, pageSize = "auto", idCols = TRUE, combine = 1L, ...,
  MoreArgs = NULL)
}
\arguments{
\item{endpoint}{Name of an endpoint, such as 'companies', 'securities', 'indices', 'owners' or 'stock_exchanges'.
See \href{docs.intrinio.com}{API documentation}. May have a composite name, e.g. \code{'fundamentals/standardized'}}

\item{pageSize}{Page size for results. If set to \code{'auto'} will use the largest size allowed for the endpoint.
See \href{http://docs.intrinio.com/#paging-limits}{Paging Limits} or \code{intrOptions()$maxPageSize}}

\item{idCols}{A single-element logical. If TRUE, will add a vectors from (...)
to the resulting data. FALSE is typically needed when the data loaded already contains ID columns}

\item{combine}{A single-element integer no less than 1.
In \href{http://docs.intrinio.com/#company-sec-filings}{some cases}
it is allowed to request multiple tickers at once to save API credits. If \code{combine} is set to 10,
data will be requested in a chunks of size 10.}

\item{...}{Arguments to iterate over similar to how \code{mapply} does it.
Must be vectors of the same length with names corresponding to Intrinio API queries}

\item{MoreArgs}{A named list of other constant queries to pass to \code{intrCall}.
Elements of MoreArgs must be vectors of length 1}
}
\description{
Use if needed to load data that requires multiple requests, such as data for multiple securities.
Each request can be multipage
}

