% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intrinio.R
\name{intrOptions}
\alias{intrOptions}
\title{Gets or sets Intrinio package options}
\usage{
intrOptions(...)
}
\arguments{
\item{...}{Named arguments for option setting. Will return all current options if none are specified.}
}
\value{
A list of options if none are passed
}
\description{
To avoid passing frequent options to every API wrapper, they are defined globally (within package scope).
}
\details{
Currenly the following options are available:
\describe{
  \item{\code{verbose:}}{Will show status messages if set to \code{TRUE}.\cr
  Default: \code{TRUE}}

  \item{\code{costWarning:}}{If \code{TRUE}, will ask for user confirmation in case when request costs
  more than \code{warnThreshold} API call credits (useful for multipage calls).\cr
  Default: \code{FALSE}}

  \item{\code{warnThreshold:}}{Minimum number of API calls in multipage request to request confirmation.\cr
  Default: \code{50}}

  \item{\code{outFormat}:}{Format of resulting output.\cr
  Possible options: \code{'data.frame'} or \code{'data.table'}.\cr
  Default: \code{'data.frame'}}

  \item{\code{maxPageSize} (view only):}{
  A \code{data.frame} with maximum page size by endpoint.
  Used to determine default page size in \code{intrMultipage} call.
  To override, pass \code{page_size} to API call when needed. See \href{http://docs.intrinio.com/#paging-limits}{Paging Limits}
  }
}
}
\note{
Options are reset to default on every package reload.
}
\examples{
# get option list
intrOptions()

# modify options
intrOptions(verbose = TRUE, costWarning = FALSE)
}
