% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postProcessing.R
\name{intrRbind}
\alias{intrRbind}
\title{Rbinds raw list data into tables}
\usage{
intrRbind(x, id = NULL, outFormat = intrOptions()$outFormat)
}
\arguments{
\item{x}{a list with raw data}

\item{id}{A vector or list with id columns. Use only together with \code{intrCallWrap}.}
}
\value{
returns data in specified format, determined with \code{\link{intrOptions}}
}
\description{
Rbinds raw list data into tables
}

