% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intrinio.R
\name{intrAuth}
\alias{intrAuth}
\title{Authorize Web API access}
\usage{
intrAuth(user, pass)
}
\arguments{
\item{user}{A string containing \code{API_USERNAME}}

\item{pass}{A string containing \code{API_PASSWORD}}
}
\value{
a list with credentials if user and password are not passed.
}
\description{
The Intrinio API uses Basic Authentication over HTTPS.
You can find your \code{API_USERNAME} and \code{API_PASSWORD} on your User Admin Dashboard of the site.\cr
\code{intrAuth} simply stores your credentials in \code{intrinio_globals} environment for future use.
Credentials need to be specified on every package reload.
}
\examples{
intrAuth('a543b029ec335ddfw66dd95bfa1ea3ac', '991d8ca963fgaa357aef78b4784d88b0')
intrAuth()
}
\seealso{
\href{http://blog.intrinio.com/getting-started-with-intrinio/}{Getting Started article}
}

